Extract Tabs to their own component
